Change root vue to default.vue for future functional expansion.
