+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reindeer-editor",
-  "version": "0.2.3",
+  "version": "0.2.4",
   "description": "A documentation generation tool supported by Reindeer Project. You can edit a CDD and display the cloud computing architecture by using this.",
   "author": "ShintaroHara",
   "private": true,
